% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACQ4_functions.R
\name{downsample}
\alias{downsample}
\alias{downsample.default}
\alias{downsample.ts}
\title{Downsample a signal by an integer factor}
\usage{
downsample(x, n, phase = 0)

\method{downsample}{default}(x, n, phase = 0)

\method{downsample}{ts}(x, n, phase = 0)
}
\arguments{
\item{x}{input data, specified as a numeric vector or matrix. In case of a
vector it represents a single signal; in case of a matrix each column is a
signal.}

\item{n}{downsampling factor, specified as a positive integer.}

\item{phase}{offset, specified as a positive integer from \code{0} to \code{n
- 1}. Default: 0.}
}
\description{
Downsample a signal by an integer factor
}
\examples{
x=1:10
x
downsample(x, 2)
x3=matrix(1:30, ncol=3, dimnames=list(NULL, c("x", "y", "z")))
x3
downsample(x3, 2)
xts=stats::ts(1:10, start=0, deltat=0.1)
xts
downsample(xts, 2)

x3ts=stats::ts(matrix(1:30, ncol=3, dimnames=list(NULL, c("x", "y", "z"))), start=0, deltat=0.1)
x3ts
downsample(x3ts, 2)
}
