% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACQ4_functions.R
\name{downsample_ts}
\alias{downsample_ts}
\title{Downsample a signal by an integer factor}
\usage{
downsample_ts(x, n, phase = 0)
}
\arguments{
\item{x}{input data, specified as a numeric vector or matrix. In case of a
vector it represents a single signal; in case of a matrix each column is a
signal.}

\item{n}{downsampling factor, specified as a positive integer.}

\item{phase}{offset, specified as a positive integer from \code{0} to \code{n
- 1}. Default: 0.}
}
\description{
Downsample a signal by an integer factor
}
\examples{
xts=stats::ts(1:10, start=0, deltat=0.1)
xts
downsample_ts(xts, 2)
}
