% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACQ4_functions.R
\name{h52ts}
\alias{h52ts}
\title{Read data from an ACQ4 HDF5 file and turn it into an R time series}
\usage{
h52ts(files, name = "data", bit64conversion = "double", channel = 2, ...)
}
\arguments{
\item{...}{}
}
\value{
A \code{ts} or \code{mts} object
}
\description{
Read data from an ACQ4 HDF5 file and turn it into an R time series
}
