% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racq4-package.R
\docType{package}
\name{racq4-package}
\alias{racq4}
\alias{racq4-package}
\title{racq4: Analysis of acq4 electrophysiology data}
\description{
Functions to analyse acq4 electrophysiology data in its native HDF5 format including, reading (with attributes), downsampling, smoothing, spike finding and plotting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jefferislab/racq4}
  \item \url{https://jefferislab.github.io/racq4/}
  \item Report bugs at \url{https://github.com/jefferislab/racq4/issues}
}

}
\author{
\strong{Maintainer}: Shahar Frechter \email{frechter@mrc-lmb.cam.ac.uk} (\href{https://orcid.org/0000-0002-0431-5849}{ORCID})

Authors:
\itemize{
  \item Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})
}

}
\keyword{internal}
