% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACQ4_functions.R
\name{spike_times_folder}
\alias{spike_times_folder}
\title{Read and analyse spikes in all experiments of a single type for a single cell}
\usage{
spike_times_folder(base_folder, experiment_type, file_extension, threshold)
}
\arguments{
\item{threshold}{}
}
\description{
Read and analyse spikes in all experiments of a single type for a single cell
}
\examples{
\dontrun{
}
}
