% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACQ4_functions.R
\name{spikes}
\alias{spikes}
\title{Read and analyse spikes in a single ACQ4 experiment}
\usage{
spikes(base_folder, query_folder, file_extension, threshold)
}
\arguments{
\item{threshold}{}
}
\description{
Read and analyse spikes in a single ACQ4 experiment
}
\examples{
\dontrun{
}
}
