% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racq4-package.R
\docType{package}
\name{racq4-package}
\alias{racq4}
\alias{racq4-package}
\title{racq4: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jefferislab/racq4}
  \item \url{https://jefferislab.github.io/racq4/}
  \item Report bugs at \url{https://github.com/jefferislab/racq4/issues}
}

}
\keyword{internal}
